# backend/api_contract.yaml
openapi: 3.0.3
info:
  title: Mastermind API
  version: 1.0.0
  description: API para jogar Mastermind (CRUD de jogo + palpites)
servers:
  - url: http://localhost:8000
    description: Servidor local
paths:
  /games:
    post:
      summary: Iniciar novo jogo
      description: Cria uma nova instância de jogo Mastermind com código secreto aleatório.
      responses:
        '200':
          description: Novo jogo criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GamePublic'
              examples:
                exemplo:
                  value:
                    id: "123e4567-e89b-12d3-a456-4ac26614174000"
                    attempts: 0
                    max_attempts: 10
                    status: active
  /games/{game_id}/guess:
    post:
      summary: Fazer um palpite
      description: Avalia as cores enviadas e retorna acertos na posição correta e acertos em posição errada.
      parameters:
        - in: path
          name: game_id
          schema:
            type: string
          required: true
          description: ID do jogo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GuessRequest'
            examples:
              exemplo:
                value:
                  colors: ["red", "blue", "green", "yellow"]
      responses:
        '200':
          description: Resultado do palpite
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuessResponse'
              examples:
                sucesso:
                  value:
                    correct_position: 1
                    wrong_position: 2
                    status: active
                    message: "Tentativa 1/10"
        '404':
          description: Jogo não encontrado ou finalizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                nao_encontrado:
                  value:
                    detail: "Jogo não encontrado ou já finalizado."
  /games/{game_id}:
    get:
      summary: Consultar status do jogo
      description: Retorna o estado atual do jogo, incluindo tentativas e status.
      parameters:
        - in: path
          name: game_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Jogo encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GamePublic'
              examples:
                exemplo:
                  value:
                    id: "123e4567-e89b-12d3-a456-426614174000"
                    attempts: 1
                    max_attempts: 10
                    status: active
        '404':
          description: Jogo não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                nao_encontrado:
                  value:
                    detail: "Jogo não encontrado"
components:
  schemas:
    Game:
      type: object
      properties:
        id:
          type: string
          format: uuid
        secret_code:
          type: array
          items:
            type: string
            enum: [red, green, blue, yellow, orange, purple]
          minItems: 4
          maxItems: 4
        attempts:
          type: integer
          example: 0
        max_attempts:
          type: integer
          example: 10
        status:
          type: string
          enum: [active, won, lost]
    GamePublic:
      type: object
      properties:
        id:
          type: string
          format: uuid
        attempts:
          type: integer
          example: 0
        max_attempts:
          type: integer
          example: 10
        status:
          type: string
          enum: [active, won, lost]
      description: Versão pública do jogo, sem o código secreto
    GuessRequest:
      type: object
      properties:
        colors:
          type: array
          items:
            type: string
            enum: [red, green, blue, yellow, orange, purple]
          minItems: 4
          maxItems: 4
      required: [colors]
    GuessResponse:
      type: object
      properties:
        correct_position:
          type: integer
          example: 1
        wrong_position:
          type: integer
          example: 2
        status:
          type: string
          enum: [active, won, lost]
        message:
          type: string
          example: "Tentativa 1/10"
      required: [correct_position, wrong_position, status, message]
    Error:
      type: object
      properties:
        detail:
          type: string